rename cohere tool parser
